More responsive controls
In the old implementation, a keyup always stopped the paddle. This
is problematic in the following scenario:

1. User presses DOWN, paddle starts going down
2. User presses UP, paddle starts going up
3. User lifts the finger off DOWN, the paddle stops

Lifting finger off DOWN should not have any effect when the paddle
is going up. This commit fixes the issue by checking what the
current direction of paddle is. Lifting finger off DOWN only cancels
the DOWN motion and similarly for the UP key.